requirements/local.txt
# Core Dependencies - Versiones validadas funcionando
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langsmith>=0.1.0,<0.2.0

# OpenAI API - Versión actual validada
openai>=1.100.0,<2.0.0

# Vector Store and Embeddings - Versiones estables funcionando
chromadb>=1.0.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0

# Database and Storage - Versiones validadas Python 3.13
# sqlite3 es built-in en Python
sqlalchemy>=2.0.0,<3.0.0
pandas>=2.2.0,<2.3.0  # Versión validada funcionando
openpyxl>=3.1.0,<4.0.0

# Web Interfaces
streamlit==1.29.0
streamlit-chat==0.1.1
streamlit-authenticator==0.2.3

# Image Processing and Vision
pillow==10.2.0
opencv-python==4.9.0.80

# API Integration
requests==2.31.0
flask>=3.1.0,<4.0.0
werkzeug>=3.0.1,<4.0.0

# Data Processing
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tenacity==8.2.3
tqdm==4.66.1

# Monitoring and Logging
structlog==23.2.0
rich==13.7.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Document Processing
pypdf==3.17.4
python-multipart==0.0.6
